services/processor_service/requirements.txt
Django>=4.2,<5
gunicorn
psycopg2-binary
celery[redis]
redis
python-dotenv
djangorestframework==3.16.1